Add: Support using a pkg repo for gitlab
The recipe used to assume that we'd want to get the gitlab package
from the S3 download location.  However it may be the case that a
customer already has staged the gitlab package in an internal repo
such as Satellite, and/or that there is not open access to download
arbitrary files from S3.

This change introduces an attribute which if set to true will cause
the remote file resource to be skipped and the source parameter in
the package resource not to be used, thus installing from the repo.